Add HP component and UI binding for Enemy
Introduced UMyActorComponent to manage Enemy HP and refactored Enemy to use this component instead of a direct HP variable. Added UMyUserWidget for HP bar UI and bound it to the HP component. Updated EnemyAnimInstance and related classes to use the new HP system. Registered UMG as a dependency in build files.

+++ Source/UnrealCpp/AI/EnemyAnimInstance.h
@@ -20,7 +20,7 @@ public:
 	UPROPERTY(Category = "Ani", EditAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
 	UAnimMontage* AttackMontage;
 	UPROPERTY(Category = "Ani", EditAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
-	float EnemyHP;
+	int32 EnemyHP;
 	UPROPERTY(Category = "Ani", EditAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
 	class UCharacterMovementComponent* CharacterMovement;
 	UPROPERTY(Category = "Ani", EditAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
